requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.9.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
pypdf>=4.0.0
